job/requirements.txt
requests==2.23.0
ipdb==0.13.2
pymongo==3.10.1
python-dateutil==2.8.1
